rename components for consistency
plural FoosTable.svelte -> singular FooTable.svelte
