+++ src/lib/explorer/ExchangeTable.svelte
