+++ src/lib/explorer/PairTable.svelte
