+++ src/routes/trading-view/[chain]/exchanges/+page.svelte
@@ -6,7 +6,7 @@
 	import { page } from '$app/stores';
 	import { goto } from '$app/navigation';
 	import Breadcrumbs from '$lib/breadcrumb/Breadcrumbs.svelte';
-	import ExchangesTable from '$lib/explorer/ExchangesTable.svelte';
+	import ExchangeTable from '$lib/explorer/ExchangeTable.svelte';
 	import { HeroBanner, Section } from '$lib/components';
 
 	export let data;
@@ -19,7 +19,7 @@
 		direction: q.get('direction') || 'desc'
 	};
 
-	async function handleChange({ detail }: ComponentEvents<ExchangesTable>['change']) {
+	async function handleChange({ detail }: ComponentEvents<ExchangeTable>['change']) {
 		await goto('?' + new URLSearchParams(detail.params), { noScroll: true });
 		detail.scrollToTop();
 	}
@@ -43,6 +43,6 @@
 	</Section>
 
 	<Section padding="sm">
-		<ExchangesTable rows={exchanges} {...options} hiddenColumns={['chain_name']} on:change={handleChange} />
+		<ExchangeTable rows={exchanges} {...options} hiddenColumns={['chain_name']} on:change={handleChange} />
 	</Section>
 </main>
