+++ src/routes/trading-view/[chain]/tokens/[token]/+page.svelte
@@ -8,7 +8,7 @@
 	import Breadcrumbs from '$lib/breadcrumb/Breadcrumbs.svelte';
 	import InfoTable from './InfoTable.svelte';
 	import InfoSummary from './InfoSummary.svelte';
-	import PairsTable from '$lib/explorer/PairsTable.svelte';
+	import PairTable from '$lib/explorer/PairTable.svelte';
 
 	export let data;
 	$: ({ token } = data);
@@ -27,7 +27,7 @@
 			...Object.fromEntries($page.url.searchParams.entries())
 		});
 
-	async function handlePairsChange({ detail }: ComponentEvents<PairsTable>['change']) {
+	async function handlePairsChange({ detail }: ComponentEvents<PairTable>['change']) {
 		await goto('?' + new URLSearchParams(detail.params), { noScroll: true });
 		detail.scrollToTop();
 	}
@@ -57,7 +57,7 @@
 		<h2>Trading pairs</h2>
 
 		{#if !$pairsClient.error}
-			<PairsTable {...$pairsClient} on:change={handlePairsChange} />
+			<PairTable {...$pairsClient} on:change={handlePairsChange} />
 		{:else}
 			<Alert>
 				An error occurred loading the pairs data. Check the URL parameters for errors and try reloading the page.
