+++ src/routes/trading-view/trading-pairs/+page.svelte
@@ -5,7 +5,7 @@
 	import type { ComponentEvents } from 'svelte';
 	import { goto } from '$app/navigation';
 	import Breadcrumbs from '$lib/breadcrumb/Breadcrumbs.svelte';
-	import PairsTable from '$lib/explorer/PairsTable.svelte';
+	import PairTable from '$lib/explorer/PairTable.svelte';
 	import { HeroBanner, Section } from '$lib/components';
 
 	export let data;
@@ -13,7 +13,7 @@
 
 	let loading = false;
 
-	async function handleChange({ detail }: ComponentEvents<PairsTable>['change']) {
+	async function handleChange({ detail }: ComponentEvents<PairTable>['change']) {
 		loading = true;
 		await goto('?' + new URLSearchParams(detail.params), { noScroll: true });
 		loading = false;
@@ -34,6 +34,6 @@
 	</Section>
 
 	<Section padding="sm">
-		<PairsTable {...pairs} {...options} {loading} on:change={handleChange} />
+		<PairTable {...pairs} {...options} {loading} on:change={handleChange} />
 	</Section>
 </main>
